% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{columnSums}
\alias{columnSums}
\title{Simple RcppSparse::Matrix example}
\usage{
columnSums(A)
}
\arguments{
\item{A}{an object of class \code{dgCMatrix}}
}
\description{
Compute column sums of a sparse matrix
}
\examples{
library(Matrix)
A <- rsparsematrix(nrow = 10, ncol = 5, density = 0.5)
columnSums(A)

# Relevant C++ code:
\dontrun{
```{Rcpp}
Rcpp::NumericVector columnSums(RcppSparse::Matrix& A) {
     Rcpp::NumericVector sums(A.cols());
     for (size_t col = 0; col < A.cols(); ++col)
          for (RcppSparse::Matrix::InnerIterator it(A, col); it; ++it)
               sums(col) += it.value();
     return sums;
}
```
}
}
